Refactor `Depl`
